+++ app/tests/imports/test_ets_import.py
@@ -1,0 +1,4 @@
+from django.test import TestCase
+
+class EtsImporterTest(TestCase):
+
