+++ app/tests/imports/test_pa_import.py
@@ -1,0 +1,3 @@
+from django.test import TestCase
+
+class EtsImporterTest(TestCase):
