+++ app/tests/imports/utils/mock_api.py
@@ -1,0 +1,32 @@
+def generate_mock_api(m):
+        url_dict = {
+            'https://www.itis.gov/ITISWebService/jsonservice/getTaxonomicRankNameFromTSN?tsn=18402' : '{"class":"gov.usgs.itis.itis_service.data.SvcTaxonRankInfo","kingdomId":"3","kingdomName":"Plantae","rankId":"180","rankName":"Genus","tsn":"18402"}',
+            'http://www.itis.gov/ITISWebService/jsonservice/getITISTermsFromScientificName?srchKey=Mangifera':'{"class":"gov.usgs.itis.itis_service.data.SvcItisTermList","itisTerms":[{"author":"L.","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":["mango"],"nameUsage":"accepted","scientificName":"Mangifera","tsn":"28802"},{"author":"L.","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":["mango"],"nameUsage":"accepted","scientificName":"Mangifera indica","tsn":"28803"},{"author":"(Stebbing, 1914)","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":["mango bark beetle"],"nameUsage":"valid","scientificName":"Hypocryphalus mangiferae","tsn":"114931"},{"author":"(Fabricius, 1775)","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"valid","scientificName":"Sternochetus mangiferae","tsn":"188097"},{"author":"(Green, 1889)","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"valid","scientificName":"Protopulvinaria mangiferae","tsn":"200874"},{"author":"Blanco","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":["pahutan"],"nameUsage":"accepted","scientificName":"Mangifera altissima","tsn":"506471"},{"author":"Jack","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":["binjai"],"nameUsage":"accepted","scientificName":"Mangifera caesia","tsn":"506472"},{"author":"Lour.","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":["bachang"],"nameUsage":"accepted","scientificName":"Mangifera foetida","tsn":"506473"},{"author":"Griff.","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":["kwini"],"nameUsage":"accepted","scientificName":"Mangifera odorata","tsn":"506474"},{"author":"Cotes, 1893","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"invalid","scientificName":"Dacus ferrugineus var. mangiferae","tsn":"676634"},{"author":"Austin, 1984","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"valid","scientificName":"Synopeas mangiferae","tsn":"733051"},{"author":"Lesne, 1921","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"valid","scientificName":"Dinoderus mangiferae","tsn":"817533"},{"author":"Chûjô, 1936","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"valid","scientificName":"Sinoxylon mangiferae","tsn":"817723"},{"author":"Workman, 1896","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"valid","scientificName":"Cheiracanthium mangiferae","tsn":"875934"}],"requestedName":"Mangifera"}',
+            'https://www.itis.gov/ITISWebService/jsonservice/getFullHierarchyFromTSN?tsn=28802':'{"author":"","class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecordList","hierarchyList":[{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"","parentTsn":"","rankName":"Kingdom","taxonName":"Plantae","tsn":"202422"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Plantae","parentTsn":"202422","rankName":"Subkingdom","taxonName":"Viridiplantae","tsn":"954898"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Viridiplantae","parentTsn":"954898","rankName":"Infrakingdom","taxonName":"Streptophyta","tsn":"846494"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Streptophyta","parentTsn":"846494","rankName":"Superdivision","taxonName":"Embryophyta","tsn":"954900"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Embryophyta","parentTsn":"954900","rankName":"Division","taxonName":"Tracheophyta","tsn":"846496"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Tracheophyta","parentTsn":"846496","rankName":"Subdivision","taxonName":"Spermatophytina","tsn":"846504"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Spermatophytina","parentTsn":"846504","rankName":"Class","taxonName":"Magnoliopsida","tsn":"18063"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Magnoliopsida","parentTsn":"18063","rankName":"Superorder","taxonName":"Rosanae","tsn":"846548"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Rosanae","parentTsn":"846548","rankName":"Order","taxonName":"Sapindales","tsn":"28643"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Sapindales","parentTsn":"28643","rankName":"Family","taxonName":"Anacardiaceae","tsn":"28771"},{"author":"L.","class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Anacardiaceae","parentTsn":"28771","rankName":"Genus","taxonName":"Mangifera","tsn":"28802"},{"author":"Blanco","class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Mangifera","parentTsn":"28802","rankName":"Species","taxonName":"Mangifera altissima","tsn":"506471"},{"author":"Jack","class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Mangifera","parentTsn":"28802","rankName":"Species","taxonName":"Mangifera caesia","tsn":"506472"},{"author":"Lour.","class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Mangifera","parentTsn":"28802","rankName":"Species","taxonName":"Mangifera foetida","tsn":"506473"},{"author":"L.","class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Mangifera","parentTsn":"28802","rankName":"Species","taxonName":"Mangifera indica","tsn":"28803"},{"author":"Griff.","class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Mangifera","parentTsn":"28802","rankName":"Species","taxonName":"Mangifera odorata","tsn":"506474"}],"rankName":"","sciName":"","tsn":"28802"}',
+            'https://www.itis.gov/ITISWebService/jsonservice/getTaxonomicRankNameFromTSN?tsn=28802':'{"class":"gov.usgs.itis.itis_service.data.SvcTaxonRankInfo","kingdomId":"3","kingdomName":"Plantae","rankId":"180","rankName":"Genus","tsn":"28802"}',
+            'http://www.itis.gov/ITISWebService/jsonservice/getITISTermsFromScientificName?srchKey=Rhynchophorus':'{"class":"gov.usgs.itis.itis_service.data.SvcItisTermList","itisTerms":[{"author":"Herbst, 1795","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"valid","scientificName":"Rhynchophorus","tsn":"114819"},{"author":"(Linnaeus, 1758)","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"valid","scientificName":"Rhynchophorus palmarum","tsn":"114820"},{"author":"(Fabricius, 1775)","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"valid","scientificName":"Rhynchophorus cruentatus","tsn":"618909"},{"author":"Say, 1831","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"invalid","scientificName":"Rhynchophorus cicatricosus","tsn":"620082"},{"author":"Erichson, 1847","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"invalid","scientificName":"Rhynchophorus cycadis","tsn":"620083"},{"author":"Chevrolat, 1880","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"invalid","scientificName":"Rhynchophorus depressus","tsn":"620084"},{"author":"Say, 1831","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"invalid","scientificName":"Rhynchophorus immunis","tsn":"620085"},{"author":"Chevrolat, 1880","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"invalid","scientificName":"Rhynchophorus lanuginosus","tsn":"620086"},{"author":"Say, 1831","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"invalid","scientificName":"Rhynchophorus placidus","tsn":"620087"},{"author":"Say, 1831","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"invalid","scientificName":"Rhynchophorus praepotens","tsn":"620088"},{"author":"Say, 1831","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"invalid","scientificName":"Rhynchophorus truncatus","tsn":"620089"},{"author":"Fahraeus, 1845","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"invalid","scientificName":"Rhynchophorus zimmermanni","tsn":"620090"},{"author":"(Boulenger, 1898)","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"valid","scientificName":"Campylomormyrus rhynchophorus","tsn":"649829"},{"author":"Fain, 1974","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"valid","scientificName":"Curculanoetus rhynchophorus","tsn":"1188219"}],"requestedName":"Rhynchophorus"}',
+            'https://www.itis.gov/ITISWebService/jsonservice/getFullHierarchyFromTSN?tsn=114819':'{"author":"","class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecordList","hierarchyList":[{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"","parentTsn":"","rankName":"Kingdom","taxonName":"Animalia","tsn":"202423"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Animalia","parentTsn":"202423","rankName":"Subkingdom","taxonName":"Bilateria","tsn":"914154"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Bilateria","parentTsn":"914154","rankName":"Infrakingdom","taxonName":"Protostomia","tsn":"914155"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Protostomia","parentTsn":"914155","rankName":"Superphylum","taxonName":"Ecdysozoa","tsn":"914158"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Ecdysozoa","parentTsn":"914158","rankName":"Phylum","taxonName":"Arthropoda","tsn":"82696"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Arthropoda","parentTsn":"82696","rankName":"Subphylum","taxonName":"Hexapoda","tsn":"563886"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Hexapoda","parentTsn":"563886","rankName":"Class","taxonName":"Insecta","tsn":"99208"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Insecta","parentTsn":"99208","rankName":"Subclass","taxonName":"Pterygota","tsn":"100500"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Pterygota","parentTsn":"100500","rankName":"Infraclass","taxonName":"Neoptera","tsn":"563890"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Neoptera","parentTsn":"563890","rankName":"Superorder","taxonName":"Holometabola","tsn":"914213"},{"author":"Linnaeus, 1758","class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Holometabola","parentTsn":"914213","rankName":"Order","taxonName":"Coleoptera","tsn":"109216"},{"author":"Emery, 1886","class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Coleoptera","parentTsn":"109216","rankName":"Suborder","taxonName":"Polyphaga","tsn":"112747"},{"author":"Lameere, 1938","class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Polyphaga","parentTsn":"112747","rankName":"Infraorder","taxonName":"Cucujiformia","tsn":"678305"},{"author":"Latreille, 1802","class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Cucujiformia","parentTsn":"678305","rankName":"Superfamily","taxonName":"Curculionoidea","tsn":"114654"},{"author":"Latreille, 1802","class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Curculionoidea","parentTsn":"114654","rankName":"Family","taxonName":"Curculionidae","tsn":"114666"},{"author":"Schönherr, 1825","class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Curculionidae","parentTsn":"114666","rankName":"Subfamily","taxonName":"Dryophthorinae","tsn":"678828"},{"author":"Herbst, 1795","class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Dryophthorinae","parentTsn":"678828","rankName":"Genus","taxonName":"Rhynchophorus","tsn":"114819"},{"author":"(Fabricius, 1775)","class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Rhynchophorus","parentTsn":"114819","rankName":"Species","taxonName":"Rhynchophorus cruentatus","tsn":"618909"},{"author":"(Linnaeus, 1758)","class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Rhynchophorus","parentTsn":"114819","rankName":"Species","taxonName":"Rhynchophorus palmarum","tsn":"114820"}],"rankName":"","sciName":"","tsn":"114819"}',
+            'https://www.itis.gov/ITISWebService/jsonservice/getTaxonomicRankNameFromTSN?tsn=114819':'{"class":"gov.usgs.itis.itis_service.data.SvcTaxonRankInfo","kingdomId":"5","kingdomName":"Animalia","rankId":"180","rankName":"Genus","tsn":"114819"}',
+            'http://www.itis.gov/ITISWebService/jsonservice/getITISTermsFromScientificName?srchKey=Grasshoppers':'{"class":"gov.usgs.itis.itis_service.data.SvcItisTermList","itisTerms":[{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":["grasshoppers"],"nameUsage":"valid","scientificName":"Orthoptera","tsn":"102160"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":["pygmy grasshoppers"],"nameUsage":"valid","scientificName":"Tetrigidae","tsn":"102162"},{"author":"Burr, 1898","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":["monkey grasshoppers"],"nameUsage":"valid","scientificName":"Eumastacidae","tsn":"102191"},{"author":"MacLeay, 1819","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":["grasshoppers","short-horned grasshoppers"],"nameUsage":"valid","scientificName":"Acrididae","tsn":"102195"},{"author":"Krauss, 1902","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":["long-horned grasshoppers"],"nameUsage":"valid","scientificName":"Tettigoniidae","tsn":"102232"},{"author":"Thunberg, 1815","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":["smaller meadow grasshoppers"],"nameUsage":"valid","scientificName":"Conocephalus","tsn":"102249"},{"author":"Blanchard, 1845","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":["wingless long-horned grasshoppers"],"nameUsage":"valid","scientificName":"Gryllacrididae","tsn":"102278"}],"requestedName":"Grasshoppers"}',
+            'http://www.itis.gov/ITISWebService/jsonservice/getITISTermsFromScientificName?srchKey=Grasshoppers%20gregaria' : '{"class":"gov.usgs.itis.itis_service.data.SvcItisTermList","itisTerms":[null],"requestedName":"Grasshoppers gregaria"}',
+            'http://www.itis.gov/ITISWebService/jsonservice/getITISTermsFromScientificName?srchKey=Taraxacum%20officinale':'{"class":"gov.usgs.itis.itis_service.data.SvcItisTermList","itisTerms":[{"author":"F.H. Wigg.","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":["dandelion","blowball","faceclock","common dandelion"],"nameUsage":"accepted","scientificName":"Taraxacum officinale","tsn":"36213"},{"author":"(Ledeb.) Schinz ex Thell.","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":["common dandelion","rough dandelion","fleshy dandelion","horned dandelion"],"nameUsage":"not accepted","scientificName":"Taraxacum officinale ssp. ceratophorum","tsn":"524741"},{"author":"F.H. Wigg.","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":["wandering dandelion","common dandelion","lesser hawkbit"],"nameUsage":"accepted","scientificName":"Taraxacum officinale ssp. officinale","tsn":"524742"},{"author":"(Lam.) Schinz & R. Keller","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":["common dandelion"],"nameUsage":"not accepted","scientificName":"Taraxacum officinale ssp. vulgare","tsn":"524743"},{"author":"(Lyons) Blytt","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"not accepted","scientificName":"Taraxacum officinale var. palustre","tsn":"541136"}],"requestedName":"Taraxacum officinale"}',
+            'https://www.itis.gov/ITISWebService/jsonservice/getFullHierarchyFromTSN?tsn=36213':'{"author":"","class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecordList","hierarchyList":[{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"","parentTsn":"","rankName":"Kingdom","taxonName":"Plantae","tsn":"202422"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Plantae","parentTsn":"202422","rankName":"Subkingdom","taxonName":"Viridiplantae","tsn":"954898"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Viridiplantae","parentTsn":"954898","rankName":"Infrakingdom","taxonName":"Streptophyta","tsn":"846494"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Streptophyta","parentTsn":"846494","rankName":"Superdivision","taxonName":"Embryophyta","tsn":"954900"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Embryophyta","parentTsn":"954900","rankName":"Division","taxonName":"Tracheophyta","tsn":"846496"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Tracheophyta","parentTsn":"846496","rankName":"Subdivision","taxonName":"Spermatophytina","tsn":"846504"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Spermatophytina","parentTsn":"846504","rankName":"Class","taxonName":"Magnoliopsida","tsn":"18063"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Magnoliopsida","parentTsn":"18063","rankName":"Superorder","taxonName":"Asteranae","tsn":"846535"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Asteranae","parentTsn":"846535","rankName":"Order","taxonName":"Asterales","tsn":"35419"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Asterales","parentTsn":"35419","rankName":"Family","taxonName":"Asteraceae","tsn":"35420"},{"author":"F.H. Wigg.","class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Asteraceae","parentTsn":"35420","rankName":"Genus","taxonName":"Taraxacum","tsn":"36199"},{"author":"F.H. Wigg.","class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Taraxacum","parentTsn":"36199","rankName":"Species","taxonName":"Taraxacum officinale","tsn":"36213"},{"author":"F.H. Wigg.","class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Taraxacum officinale","parentTsn":"36213","rankName":"Subspecies","taxonName":"Taraxacum officinale ssp. officinale","tsn":"524742"}],"rankName":"","sciName":"","tsn":"36213"}',
+            'http://www.itis.gov/ITISWebService/jsonservice/getITISTermsFromScientificName?srchKey=Voikukka':'{"class":"gov.usgs.itis.itis_service.data.SvcItisTermList","itisTerms":[null],"requestedName":"Voikukka"}',
+            'http://www.itis.gov/ITISWebService/jsonservice/getITISTermsFromScientificName?srchKey=Galipea%20officinalis':'{"class":"gov.usgs.itis.itis_service.data.SvcItisTermList","itisTerms":[{"author":"Hanc.","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":["Angostura tree"],"nameUsage":"not accepted","scientificName":"Galipea officinalis","tsn":"506391"}],"requestedName":"Galipea officinalis"}',
+            'http://www.itis.gov/ITISWebService/jsonservice/getITISTermsFromScientificName?srchKey=Oxya%20japonica':'{"class":"gov.usgs.itis.itis_service.data.SvcItisTermList","itisTerms":[{"author":"(Thunberg, 1815,","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":["Japanese grasshopper"],"nameUsage":"invalid","scientificName":"Oxya japonica","tsn":"102221"}],"requestedName":"Oxya japonica"}',
+            'https://www.itis.gov/ITISWebService/jsonservice/getAcceptedNamesFromTSN?tsn=102221':'{"acceptedNames":[{"acceptedName":"Oxya velox","acceptedTsn":"650544","author":"(Fabricius, 1787,","class":"gov.usgs.itis.itis_service.data.SvcAcceptedName"}],"class":"gov.usgs.itis.itis_service.data.SvcAcceptedNameList","tsn":"102221"}',
+            'https://www.itis.gov/ITISWebService/jsonservice/getFullHierarchyFromTSN?tsn=650544':'{"author":"","class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecordList","hierarchyList":[{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"","parentTsn":"","rankName":"Kingdom","taxonName":"Animalia","tsn":"202423"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Animalia","parentTsn":"202423","rankName":"Subkingdom","taxonName":"Bilateria","tsn":"914154"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Bilateria","parentTsn":"914154","rankName":"Infrakingdom","taxonName":"Protostomia","tsn":"914155"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Protostomia","parentTsn":"914155","rankName":"Superphylum","taxonName":"Ecdysozoa","tsn":"914158"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Ecdysozoa","parentTsn":"914158","rankName":"Phylum","taxonName":"Arthropoda","tsn":"82696"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Arthropoda","parentTsn":"82696","rankName":"Subphylum","taxonName":"Hexapoda","tsn":"563886"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Hexapoda","parentTsn":"563886","rankName":"Class","taxonName":"Insecta","tsn":"99208"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Insecta","parentTsn":"99208","rankName":"Subclass","taxonName":"Pterygota","tsn":"100500"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Pterygota","parentTsn":"100500","rankName":"Infraclass","taxonName":"Neoptera","tsn":"563890"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Neoptera","parentTsn":"563890","rankName":"Superorder","taxonName":"Polyneoptera","tsn":"914215"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Polyneoptera","parentTsn":"914215","rankName":"Order","taxonName":"Orthoptera","tsn":"102160"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Orthoptera","parentTsn":"102160","rankName":"Suborder","taxonName":"Caelifera","tsn":"102161"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Caelifera","parentTsn":"102161","rankName":"Infraorder","taxonName":"Acrididea","tsn":"657454"},{"author":"MacLeay, 1819","class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Acrididea","parentTsn":"657454","rankName":"Superfamily","taxonName":"Acridoidea","tsn":"650497"},{"author":"MacLeay, 1819","class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Acridoidea","parentTsn":"650497","rankName":"Family","taxonName":"Acrididae","tsn":"102195"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Acrididae","parentTsn":"102195","rankName":"Subfamily","taxonName":"Oxyinae","tsn":"650537"},{"author":"Serville, 1831","class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Oxyinae","parentTsn":"650537","rankName":"Genus","taxonName":"Oxya","tsn":"102220"},{"author":"(Fabricius, 1787,","class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Oxya","parentTsn":"102220","rankName":"Species","taxonName":"Oxya velox","tsn":"650544"}],"rankName":"","sciName":"","tsn":"650544"}',
+            'https://www.itis.gov/ITISWebService/jsonservice/getITISTermsFromScientificName?srchKey=Taraxacum%20officinale':'{"class":"gov.usgs.itis.itis_service.data.SvcItisTermList","itisTerms":[{"author":"F.H. Wigg.","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":["dandelion","blowball","faceclock","common dandelion"],"nameUsage":"accepted","scientificName":"Taraxacum officinale","tsn":"36213"},{"author":"(Ledeb., Schinz ex Thell.","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":["common dandelion","rough dandelion","fleshy dandelion","horned dandelion"],"nameUsage":"not accepted","scientificName":"Taraxacum officinale ssp. ceratophorum","tsn":"524741"},{"author":"F.H. Wigg.","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":["wandering dandelion","common dandelion","lesser hawkbit"],"nameUsage":"accepted","scientificName":"Taraxacum officinale ssp. officinale","tsn":"524742"},{"author":"(Lam., Schinz & R. Keller","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":["common dandelion"],"nameUsage":"not accepted","scientificName":"Taraxacum officinale ssp. vulgare","tsn":"524743"},{"author":"(Lyons, Blytt","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"not accepted","scientificName":"Taraxacum officinale var. palustre","tsn":"541136"}],"requestedName":"Taraxacum officinale"}',
+            'https://www.itis.gov/ITISWebService/jsonservice/getAcceptedNamesFromTSN?tsn=61653':'{"acceptedNames":[{"acceptedName":"Marilynia macrodentata","acceptedTsn":"61652","author":"(Wieser, 1959,","class":"gov.usgs.itis.itis_service.data.SvcAcceptedName"}],"class":"gov.usgs.itis.itis_service.data.SvcAcceptedNameList","tsn":"61653"}',
+            'https://www.itis.gov/ITISWebService/jsonservice/getFullHierarchyFromTSN?tsn=61652':'{"author":"","class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecordList","hierarchyList":[{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"","parentTsn":"","rankName":"Kingdom","taxonName":"Animalia","tsn":"202423"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Animalia","parentTsn":"202423","rankName":"Subkingdom","taxonName":"Bilateria","tsn":"914154"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Bilateria","parentTsn":"914154","rankName":"Infrakingdom","taxonName":"Protostomia","tsn":"914155"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Protostomia","parentTsn":"914155","rankName":"Superphylum","taxonName":"Ecdysozoa","tsn":"914158"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Ecdysozoa","parentTsn":"914158","rankName":"Phylum","taxonName":"Nematoda","tsn":"59490"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Nematoda","parentTsn":"59490","rankName":"Class","taxonName":"Chromadorea","tsn":"914188"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Chromadorea","parentTsn":"914188","rankName":"Subclass","taxonName":"Chromadoria","tsn":"59492"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Chromadoria","parentTsn":"59492","rankName":"Order","taxonName":"Desmodorida","tsn":"60662"},{"author":"Filipjev, 1918","class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Desmodorida","parentTsn":"60662","rankName":"Family","taxonName":"Cyatholaimidae","tsn":"61479"},{"author":"Hopper, 1972","class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Cyatholaimidae","parentTsn":"61479","rankName":"Genus","taxonName":"Marilynia","tsn":"61648"},{"author":"(Wieser, 1959,","class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Marilynia","parentTsn":"61648","rankName":"Species","taxonName":"Marilynia macrodentata","tsn":"61652"}],"rankName":"","sciName":"","tsn":"61652"}',
+            'http://www.itis.gov/ITISWebService/jsonservice/getITISTermsFromScientificName?srchKey=Gregaria':'{"class":"gov.usgs.itis.itis_service.data.SvcItisTermList","itisTerms":[{"author":"Donkin","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"accepted","scientificName":"Navicula gregaria","tsn":"3860"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"accepted","scientificName":"Chlamydomonas gregaria","tsn":"5451"},{"author":"(A. Heller) McNeill","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":["brittle sandwort"],"nameUsage":"not accepted","scientificName":"Minuartia nuttallii ssp. gregaria","tsn":"20007"},{"author":"E.P. Bicknell","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"not accepted","scientificName":"Sanicula gregaria","tsn":"29853"},{"author":"(L. Agassiz, 1862)","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"valid","scientificName":"Clytia gregaria","tsn":"49604"},{"author":"G. O. Sars, 1885","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"valid","scientificName":"Thysanoessa gregaria","tsn":"95572"},{"author":"Fischer von Waldheim, 1820","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"valid","scientificName":"Nebria gregaria","tsn":"109489"},{"author":"Gagne, 1981","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"valid","scientificName":"Endaphis gregaria","tsn":"125280"},{"author":"Melander, 1902","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"valid","scientificName":"Hilara congregaria","tsn":"136067"},{"author":"(Frick, 1954)","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"valid","scientificName":"Chromatomyia gregaria","tsn":"143867"},{"author":"Frick, 1954","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"invalid","scientificName":"Phytomyza gregaria","tsn":"143868"},{"author":"(Lesson, 1830)","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"valid","scientificName":"Paramolgula gregaria","tsn":"159627"},{"author":"Lesson, 1830","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"invalid","scientificName":"Cynthia gregaria","tsn":"159628"},{"author":"(Lesson, 1830)","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"invalid","scientificName":"Molgula gregaria","tsn":"159629"},{"author":"(Heller, 1867)","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"valid","scientificName":"Gregarinidra gregaria","tsn":"201350"},{"author":"Heller, 1867","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"invalid","scientificName":"Membranipora gregaria","tsn":"201351"},{"author":"(G. O. Sars, 1892)","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"valid","scientificName":"Proboloides gregaria","tsn":"206566"},{"author":"A. Heller","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"not accepted","scientificName":"Arenaria gregaria","tsn":"509050"},{"author":"(A. Heller) Maguire","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"not accepted","scientificName":"Arenaria nuttallii ssp. gregaria","tsn":"525264"},{"author":"(A. Heller) Jeps.","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"not accepted","scientificName":"Arenaria nuttallii var. gregaria","tsn":"532241"},{"author":"(Mitt.) R.H. Zander","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"accepted","scientificName":"Barbula indica var. gregaria","tsn":"549337"},{"author":"Warncke, 1974","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"valid","scientificName":"Andrena gregaria","tsn":"752445"},{"author":"(Schrottky, 1905)","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"valid","scientificName":"Hypanthidioides gregaria","tsn":"755970"},{"author":"Pedro and Camargo, 2003","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"valid","scientificName":"Partamona gregaria","tsn":"763967"},{"author":"(Warncke, 1992)","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"valid","scientificName":"Hoplitis benoisti gregaria","tsn":"768026"},{"author":"Boiss. & Heldr.","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"not accepted","scientificName":"Vicia gregaria","tsn":"820766"},{"author":"(Boiss. & Heldr.) P.H. Davis","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"accepted","scientificName":"Vicia canescens ssp. gregaria","tsn":"820833"},{"author":"(A. Heller) Rabeler & R.L. Hartm.","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":["brittle sandwort"],"nameUsage":"accepted","scientificName":"Minuartia nuttallii var. gregaria","tsn":"823698"},{"author":"(Weise, 1905)","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"valid","scientificName":"Chiridopsis gregaria","tsn":"840272"},{"author":"(Brause) Domin","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"accepted","scientificName":"Cyathea gregaria","tsn":"914374"},{"author":"Brause","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"not accepted","scientificName":"Alsophila gregaria","tsn":"915203"},{"author":"(W. Fox, 1898)","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"valid","scientificName":"Mimesa gregaria","tsn":"1016716"},{"author":"(Malloch, 1933)","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"invalid","scientificName":"Mimesa gregaria simplex","tsn":"1018049"},{"author":"Lange-Bertalot and U. Rumrich in U. Rumrich, Lange-Bertalot and Rumrich","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"accepted","scientificName":"Navicula supergregaria","tsn":"1020505"},{"author":"Marloth","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"accepted","scientificName":"Euphorbia gregaria","tsn":"1026541"},{"author":"(Marloth) P.V. Heath","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"not accepted","scientificName":"Tirucalia gregaria","tsn":"1031577"},{"author":"Delève, 1963","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"valid","scientificName":"Pseudelmidolia gregaria","tsn":"1064962"},{"author":"Colenso","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"accepted","scientificName":"Fossombronia gregaria","tsn":"1108968"},{"author":"(Hook.f. & Taylor) Gottsche, Lindenb. & Nees","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"accepted","scientificName":"Plagiochila gregaria","tsn":"1113513"},{"author":"Geoffroy, 1762","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"invalid","scientificName":"Apis gregaria","tsn":"1128082"},{"author":"Mitt.","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"not accepted","scientificName":"Tortula gregaria","tsn":"1136058"},{"author":"(Schrank, 1781)","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"valid","scientificName":"Microgaster gregaria","tsn":"1144346"},{"author":"Warncke, 1992","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"invalid","scientificName":"Osmia benoisti gregaria","tsn":"1165582"},{"author":"(Fabricius, 1793)","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"invalid","scientificName":"Munida gregaria","tsn":"1201726"},{"author":"Fabricius, 1793","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"invalid","scientificName":"Galathea gregaria","tsn":"1202153"},{"author":"(Fabricius, 1793)","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"valid","scientificName":"Grimothea gregaria","tsn":"1202377"}],"requestedName":"Gregaria"}',
+            'http://www.itis.gov/ITISWebService/jsonservice/getITISTermsFromScientificName?srchKey=Gregaria%20migratoria':'{"class":"gov.usgs.itis.itis_service.data.SvcItisTermList","itisTerms":[null],"requestedName":"Gregaria migratoria"}',
+            'http://www.itis.gov/ITISWebService/jsonservice/getITISTermsFromScientificName?srchKey=Migratoria':'{"class":"gov.usgs.itis.itis_service.data.SvcItisTermList","itisTerms":[{"author":"Hartert, 1903","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":["Yellow-billed Grosbeak","Chinese Grosbeak"],"nameUsage":"valid","scientificName":"Eophona migratoria","tsn":"559921"},{"author":"Hartert, 1903","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"valid","scientificName":"Eophona migratoria migratoria","tsn":"729555"},{"author":"Riley, 1915","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"valid","scientificName":"Eophona migratoria sowerbyi","tsn":"729556"},{"author":"Linnaeus, 1766","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"invalid","scientificName":"Columba migratoria","tsn":"1125301"},{"author":"Nadchatram and Wilson, 1969","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"valid","scientificName":"Mackiena migratoria","tsn":"1179427"},{"author":"Abo-Shnaf in Abo-Shnaf et al., 2020","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"valid","scientificName":"Blattisocius migratoriae","tsn":"1184428"}],"requestedName":"Migratoria"}',
+            'http://www.itis.gov/ITISWebService/jsonservice/getITISTermsFromScientificName?srchKey=Migratoria%20manilensis':'{"class":"gov.usgs.itis.itis_service.data.SvcItisTermList","itisTerms":[null],"requestedName":"Migratoria manilensis"}',
+            'http://www.itis.gov/ITISWebService/jsonservice/getITISTermsFromScientificName?srchKey=Manilensis':'{"class":"gov.usgs.itis.itis_service.data.SvcItisTermList","itisTerms":[{"author":"Philippi","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":["asian clam"],"nameUsage":"valid","scientificName":"Corbicula manilensis","tsn":"81386"},{"author":"(Marion de Procé, 1822)","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":["striped puffer"],"nameUsage":"valid","scientificName":"Arothron manilensis","tsn":"553408"},{"author":"Herre, 1923","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"valid","scientificName":"Ophichthus manilensis","tsn":"636119"},{"author":"Ashmead, 1905","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"valid","scientificName":"Macroteleia manilensis","tsn":"751193"},{"author":"Meyen, 1834","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"valid","scientificName":"Ardea purpurea manilensis","tsn":"824406"},{"author":"(Weise, 1910)","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"valid","scientificName":"Malayocassis manilensis","tsn":"840345"},{"author":"(Weise, 1910)","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"valid","scientificName":"Agoniella manilensis","tsn":"842163"},{"author":"Schultze, 1915","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"invalid","scientificName":"Trox manilensis","tsn":"929963"},{"author":"Rohwer, 1910","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"invalid","scientificName":"Notogonidea manilensis","tsn":"1010236"},{"author":"Marion de Procé, 1822","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"invalid","scientificName":"Tetrodon manilensis","tsn":"1054798"},{"author":"Martens, 1876","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"invalid","scientificName":"Varanus manilensis","tsn":"1204402"}],"requestedName":"Manilensis"}',
+            'http://www.itis.gov/ITISWebService/jsonservice/getITISTermsFromScientificName?srchKey=Ceratophyllum':'{"class":"gov.usgs.itis.itis_service.data.SvcItisTermList","itisTerms":[{"author":"L.","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":["hornwort"],"nameUsage":"accepted","scientificName":"Ceratophyllum","tsn":"18402"},{"author":"L.","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":["coon\'s-tail","coontail","hornwort","common hornwort","coon\'s tail"],"nameUsage":"accepted","scientificName":"Ceratophyllum demersum","tsn":"18403"},{"author":"Cham.","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":["prickly hornwort"],"nameUsage":"accepted","scientificName":"Ceratophyllum muricatum","tsn":"18404"},{"author":"L.","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":["soft hornwort"],"nameUsage":"accepted","scientificName":"Ceratophyllum submersum","tsn":"18405"},{"author":"Michx.","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":["threadfoot","podostémon à feuilles cornées","hornleaf riverweed"],"nameUsage":"accepted","scientificName":"Podostemum ceratophyllum","tsn":"27031"},{"author":"A. Gray","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":["spineless hornwort"],"nameUsage":"accepted","scientificName":"Ceratophyllum echinatum","tsn":"501366"},{"author":"Cham.","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"not accepted","scientificName":"Ceratophyllum apiculatum","tsn":"510606"},{"author":"Griseb.","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"not accepted","scientificName":"Ceratophyllum australe","tsn":"510607"},{"author":"Spruce ex K. Schum.","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"not accepted","scientificName":"Ceratophyllum cristatum","tsn":"510608"},{"author":"Fassett","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":["Florida hornwort"],"nameUsage":"not accepted","scientificName":"Ceratophyllum floridanum","tsn":"510609"},{"author":"Fassett","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"not accepted","scientificName":"Ceratophyllum llerenae","tsn":"510610"},{"author":"(Griseb.) Les","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":["prickly hornwort"],"nameUsage":"accepted","scientificName":"Ceratophyllum muricatum ssp. australe","tsn":"523836"},{"author":"auct. non (Cham.) Wilmot-Dear","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"not accepted","scientificName":"Ceratophyllum submersum ssp. muricatum","tsn":"525442"},{"author":"(Cham.) Asch.","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"not accepted","scientificName":"Ceratophyllum demersum var. apiculatum","tsn":"533317"},{"author":"(Cham.) Garcke","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"not accepted","scientificName":"Ceratophyllum demersum var. apiculatum","tsn":"533318"},{"author":"K. Schum.","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"not accepted","scientificName":"Ceratophyllum demersum var. cristatum","tsn":"533319"},{"author":"(A. Gray) A. Gray","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"not accepted","scientificName":"Ceratophyllum demersum var. echinatum","tsn":"533320"},{"author":"(A. Gray) Wilmot-Dear","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"not accepted","scientificName":"Ceratophyllum submersum var. echinatum","tsn":"533321"},{"author":"(Cham.) Wilmot-Dear","class":"gov.usgs.itis.itis_service.data.SvcItisTerm","commonNames":[null],"nameUsage":"not accepted","scientificName":"Ceratophyllum submersum ssp. muricatum","tsn":"897269"}],"requestedName":"Ceratophyllum"}',
+            'https://www.itis.gov/ITISWebService/jsonservice/getFullHierarchyFromTSN?tsn=18402':'{"author":"","class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecordList","hierarchyList":[{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"","parentTsn":"","rankName":"Kingdom","taxonName":"Plantae","tsn":"202422"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Plantae","parentTsn":"202422","rankName":"Subkingdom","taxonName":"Viridiplantae","tsn":"954898"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Viridiplantae","parentTsn":"954898","rankName":"Infrakingdom","taxonName":"Streptophyta","tsn":"846494"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Streptophyta","parentTsn":"846494","rankName":"Superdivision","taxonName":"Embryophyta","tsn":"954900"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Embryophyta","parentTsn":"954900","rankName":"Division","taxonName":"Tracheophyta","tsn":"846496"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Tracheophyta","parentTsn":"846496","rankName":"Subdivision","taxonName":"Spermatophytina","tsn":"846504"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Spermatophytina","parentTsn":"846504","rankName":"Class","taxonName":"Magnoliopsida","tsn":"18063"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Magnoliopsida","parentTsn":"18063","rankName":"Superorder","taxonName":"Ceratophyllanae","tsn":"846540"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Ceratophyllanae","parentTsn":"846540","rankName":"Order","taxonName":"Ceratophyllales","tsn":"846616"},{"author":null,"class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Ceratophyllales","parentTsn":"846616","rankName":"Family","taxonName":"Ceratophyllaceae","tsn":"18401"},{"author":"L.","class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Ceratophyllaceae","parentTsn":"18401","rankName":"Genus","taxonName":"Ceratophyllum","tsn":"18402"},{"author":"L.","class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Ceratophyllum","parentTsn":"18402","rankName":"Species","taxonName":"Ceratophyllum demersum","tsn":"18403"},{"author":"A. Gray","class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Ceratophyllum","parentTsn":"18402","rankName":"Species","taxonName":"Ceratophyllum echinatum","tsn":"501366"},{"author":"Cham.","class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Ceratophyllum","parentTsn":"18402","rankName":"Species","taxonName":"Ceratophyllum muricatum","tsn":"18404"},{"author":"L.","class":"gov.usgs.itis.itis_service.data.SvcHierarchyRecord","parentName":"Ceratophyllum","parentTsn":"18402","rankName":"Species","taxonName":"Ceratophyllum submersum","tsn":"18405"}],"rankName":"","sciName":"","tsn":"18402"}',
+
+        }
+        for url, data in url_dict.items():
+            m.get(url,text=data)
